[FIX] mass_mailing_themes: restore button text colors
The button text colors were forced to the default link color while that
color should only be forced for links which are not buttons.

Note: this fix will not impact existing mail templates thanks to
transcoding. That's why it is possible to fix safely in stable versions.

opw-2360756

closes odoo/enterprise#14217

X-original-commit: 4aa1ccaf18493235cc9db1dc701589982a711844
Related: odoo/odoo#60308
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>