[FIX] marketing_automation: hide state in mailing tree view
When creating templates (mailing) from the marketing campaign activities and
open them with the available stat button, the state of the mailing is set to
`sent` which might be confusing for the end user.

Indeed state has no real use here as templates (mailings) are used through
the campaign on participants. Their specific state is irrelevant, only
participant and traces are relevant in marketing automation context.

This commit hides the `state` from mailing tree view (only when opening it
from the stat button of marketing campaing) to avoid confusion.

Task ID-2300427
PR #14065

X-original-commit: 27228b028aecb5132f17b7596f3abfc02250ef36