[FIX] marketing_automation: reset 'Unicity based on' field when changing model
When a user changes the `Target` (model) for the marketing campaign,
the field `Unicity based on` is not being reset and still showing the
field selected from previously selected model. This could lead to the
use of wrong fields for unicity.

This commit improves the behavior by resetting the `Unicity based on`
field each time the user changes the Target for a campaign.

Task ID-2300427
PR #14065

X-original-commit: 042a1c4488ea4c034d2d8032bbc6d46302214ae1