[FIX] marketing_automation: properly set the subject of mailing
Inside mailing activity, when we want to create a new mailing, we enter the
name in the m2o and click `Create and Edit` from the dropdown. Name in m2o
is not copied to the subject. For the end user this behavior is confusing as
typed string should be propagated as mailing subject.

This commit improves the behavior by copying the string typed in m2o to the
subject of the mailing.

This is due to Create and Edit behavior setting default_name in context.
Mailing model uses subject as _rec_name, and not name. We therefore have to
catch default_name from context and redirect it towards subject.

A more generic fix would be to use default_display_name and let default_get
handle it accordingly to rec_name. However for this fix we limit ourselves to
a local fix and a clean solution will be engineered soon.

Task ID-2300427
PR #14065

closes odoo/enterprise#14200

X-original-commit: b83f1384a38e8db52ce2144be13f9f4406b3761b
Related: odoo/odoo#60262
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>