[FIX] social: fix comment edition display
When editing a comment in the stream.post comments window, the comment you are
currently editing is supposed to be replaced by a text area to let you input
the new value.

This textarea is correctly displayed but the edited comment is not hidden.
Even though it does not "break" anything, it looks a bit strange.

This commit fixes the behavior by correctly hiding the comment.

Task 2209729

closes odoo/enterprise#17384

X-original-commit: c9c697f00fd962a8b0efdcb24c9d10312843f495
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: awa-odoo <awa-odoo@users.noreply.github.com>