[FIX] sale_timesheet_enterprise: Fix the stop timer on invoiced timesheet
Currenlty, once timesheet is invoice and you again start timesheet from
timer header and try to stop it it will raise the validation error that
'You can not modify the invoiced timesheet' because when we stop the
timesheet from header it will try_to_match the timesheet with previous
invoiced timesheet and going to append it.

So in this commit, if do not match with invoiced timesheet so it should
not be updated when we stop the timer from timesheet header.

closes odoo/enterprise#14808

Taskid: 2357182
X-original-commit: 47080ab15bdaa154f610b76bed6000636d48ef3d
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>