[IMP] web_map: pans the map relative to popup height
before this commit: when clicking on map sidebar opens popup and map marker
is displayed in center, panTo method was centering map based on partner
partner_latitude and partner_longitude but this sometime hides popup,
popup is not fully displayed.

after this commit: map is centered but it will consider popup height first,
from partner_latitude and partner_longitude pixels are computed and height
of popup/2 will reduced from computed pixels and then pixels are again
converted back to coordinates and then panTo method is called with that
coordinates.

task-2304982

closes odoo/enterprise#15419

X-original-commit: 40cda2d9349a257382e754d52b273cafbbba28f3
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>