[FIX] documents: add test for applying share defaults
This fix updates a comment to clarify that the custom values applied to
a document created from a message come from the share, not from the
alias defaults (even though it is normally the same thing, since the
alias defaults is synchronized from the share on update).
It also adds a test to verify that the values are copied from the share
when creating a new document from a mail attachment.
There is no change in the behavior of the application itself.

opw-2352486

closes odoo/enterprise#14175

X-original-commit: 08cafe13cf96e8bc992e5635c0b948d5a5faf13c
Signed-off-by: Nicolas Galler <nicocrm@users.noreply.github.com>
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>