[FIX] project_timesheet_synchro: remove unused cookies permission
We have been warned by the Chrome Web Store that the application
doesn't use cookies. Their security policy having changed, it is
mandatory to modify this to avoid to be deleted from the store.

Note that this extension is generated from 11.0 for historical reasons.
Some work must be done in more recent version to avoid this.

closes odoo/enterprise#14271

Task-id: 2345852
X-original-commit: 851a15557ac080c633e583e9f541fb7ce8f60a5c
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>