[FIX] l10n_mx_edi: take into account Tax 0% in cfdi document
Before this commit, the cfdi:Impuestos part of the cfdi document wasn't
created if the invoice have Tax 0%. This error occurs because in the
case of Tax 0%, total_tax_details_transferred will be 0. So the
condition to create the Impuestos part of the document wasn't fulfilled.

Now, the cfdi:Impuestos part is correctly filled with the information of
the Tax 0%

opw-2357680

closes odoo/enterprise#14278

X-original-commit: 77b497273e6e19b7994155b0537c2273fdf5f9ac
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>