[FIX] sale_subscription: Do not rely on origin so all the time
Since commit a0e01a202de78ed0e5e904477e8de024c5af95c5, we rely on the
origin sale order to fetch the shipping address (and to keep
consistency the invoice-partner_id) of the original sale order of the
subscription. This causes an issue when we change the partner_id of the
subscription afterwards.

This commit addresses the issue by only relying on the sale order
information when its partner_id is the same as the subscription.

closes odoo/enterprise#14230

X-original-commit: 51fece1e749b236e1716e5caa1778ff2c8db69c8
Signed-off-by: Simon Goffin (sig) <sig@openerp.com>
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>