[FIX] l10n_be_intrastat: exclude non vat partner on intrastat report
Behavior prior to the fix:

On a database with the BE localization installed, the Accounting
Intrastat report installed, and a BE company configured, exporting the
Intrastat report to XML with the "Partners with VAT Numbers Only" option
checked includes all invoices, even the ones without a partner with VAT
number.

Behavior after the fix:

The invoices (and refunds) associated with partners without a VAT number
are correctly excluded.

opw-2365940

closes odoo/enterprise#14232

X-original-commit: 86cd5beaf96a26b52f8abcdea695be324d724364
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>
Signed-off-by: Nicolas Galler <nicocrm@users.noreply.github.com>