[FIX] sale_subscription: filter bank based on company
- Create 2 companies A & B
- Set the `__system__` user in company B
- Create the following bank accounts for partner A:
  Account B, Sequence N, Company B
  Account A, Sequence N+1, Company A
- Create a subscription in company A, start
- Wait for the cron to create the invoice

The bank account chosen is Account B instead of Account A.

This happens because although the `company` is the correct one (A), the
record rules do not apply since the user is the superuser. Therefore,
the first account is selected, which is B in this case.

As a solution, we filter out accounts in the wrong company.

opw-2359812
opw-2366360

closes odoo/enterprise#14264

X-original-commit: 6fbe32cec5ba22ac6fdedfedc91b1ef50161d61a
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>