[FIX] web_enterprise: auto close Print menu on mobile
On desktop the ActionMenus' Print menu doesn't close automatically when
one of its item is selected. This is annoying on mobile as, due to the
small screen space, this menu can hide a huge part of the screen.

This commit fixes it by auto closing this menu, but only on mobile.

Task ID: 1879566