[FIX] account_reports: Fix total account in GL with foreign currency
The total in foreign currency has been removed by:
https://github.com/odoo/enterprise/pull/14013/commits/d9cfe2a64995f262d8b9a8897377f84c7fefd86b

However, when the currency is set on the account, all journal items using this account should
share the same currency. This commit re-adds this total in foreign currency but under the condition the account is well configurated as using a foreign currency.

closes odoo/enterprise#14475

X-original-commit: b990e108fad270e94785ecb8881aedf7c856e830
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>