[IMP] account_reports: Allows to post empty tax report
A company inactive during a period should still be able to post the tax
report for that period.

This is now possible by adding two lines set to 0 to the account move
when the report is empty.

Also fix an issue with the xml export of the l10n_be tax report which
should always contains values for either the tax grid 71 or 72.

When those two lines are empty, we'll now add the grid 71 to the xml
with 0 as a value.

task id #2317046

closes odoo/enterprise#16809

X-original-commit: d4036da02b997b24115e0211d8709d565b1f2904
Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>
Signed-off-by: Josse Colpaert <jco@openerp.com>