[IMP] pos_l10n_se: Creating module for sweden certification
Allow a shop in sweden to use a certified blackbox.
This module allows reports for the daily session.
Add pro forma orders. Pro forma ticket will be marked and send to the blackbox aswell.
Reprint ticket are sent to the blackbox too.
A hash is generated for each VAT ticket.

closes odoo/enterprise#19007

Task-id: 1986405
X-original-commit: f8b32595ae7def92a31d91b5b92be1728cea198c
Related: odoo/odoo#72142
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>