[FIX] delivery_easypost: round weight up
Behavior prior to the fix:

When shipping a package with a very small weight, the weight was rounded
down to 0 oz, causing EasyPost to refuse the request.  The rounding is
always performed at 0.1 precision, because that is what EasyPost
accepts, thus a package of less than 0.05oz would be rounded down to 0.

Behavior after the fix:

When computing the weight a bottom cap of 0.1oz was introduced.  This
will only affect shipments with very low weight.

Note:

An alternative would have been to use a rounding method of UP, which
would be similar to what other carriers plugins (eg delivery_fedex) are
using, but might adversedly affect shipments with weights very slightly
above the 1oz mark (for example 1lb1.05oz would become 1lb1.1oz)

opw-2362389

closes odoo/enterprise#14335

X-original-commit: 03d8799e6d609e6a0224bedd3f0616a9426e80af
Signed-off-by: Nicolas Galler <nicocrm@users.noreply.github.com>