[IMP] hr_payroll: Work entries analysis
New report is created, this report allows to ease the work entry
analysis.
This reporting is available with a new menu entry in
Payroll/Reporting/Work Entry Analysis.
This report is used for a pivot view. In this view, we add a measure in
Days to have the number of real days under this work entry type.
By default, the group by is work entry type, the filter is the
current month and the measure is in days.

task-2365616

closes odoo/enterprise#14477

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>