[IMP] l10n_be_hr_payroll: filter working schedule like in contract
In credit time and exit credit time wizards, the working schedules are
not filter like it is the case in the contract model.
In the contract, the working schedules are filtered according to the
company. If a working schedule contains the same company than the one in
the contract or this working schedule doesn't contains the company then
we take it.
This commit add this filter in the both wizards.

task-2314223