[FIX] website_calendar: fix default timezone assignation
This commit fixes the default timezone assignation. Since 7e8d652b99c96cac1ffc63ae37257c419024031e,
the returned timezone is directly the needed string and not an object anymore.

Calling 'name()' on returned timezone is than not necessary anymore and leads
to a crash.

This commit simply remove the call to 'name()'.

Task ID: 2333825

closes odoo/enterprise#14342

X-original-commit: 9028927da430e04485dd4ee4f148ee3c836ee59d
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>