[FIX] voip: reject call properly
This fix solves 2 encountered issues when
a call is rejected. First of all, if you call
someone which is already in an active call, the
denied answer was not correctly interpreted on the
"caller" side as the session was not cancelled. Now
a message telling that the "callee" is not available
is displayed and the session is cancelled. The second
issue appears when the sip_ignore_incomming fields was
activated. In this case, the call was always rejected.
The problem is that the call may be redirected elsewhere,
on a desk phone for e.g.. The previous solution does not
allows this case.

see: https://tools.ietf.org/html/rfc3261#section-21.6.2 for
status code

closes odoo/enterprise#14341

X-original-commit: 1481ae0dee16bdb426eed1bfbd928b94e59c2163
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>