[FIX] sale_account_taxcloud: validate taxcloud after carrier change
Configure taxcloud, 2+ delivery carrier, card payment
Go to the website shop
Add product to cart
Checkout with USA address
On the Confirm Order page, the tax is computed via TaxCloud and shown on
the right.
Pay via payment acquirer site

The Subtotal (without taxes) is sent to the Payment Acquirer instead of
the Total

This occur because when selecting delivery carrier coupon lines are
recomputed. To be on the safe side with taxcloud, taxes are excluded
before
recomputing coupon lines, but never added in this specific flow.

opw-2338103
opw-2362493
opw-2353048

closes odoo/enterprise#14351

X-original-commit: c0869dcd833c5b9cf89541a22d17f2beb864faab
Related: odoo/odoo#60570
Signed-off-by: agr-odoo <agr-odoo@users.noreply.github.com>
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>