[IMP] l10n_be_hr_payroll: remove the selection for work time rate
To create a credit time, we need to choose a calendar corresponding to
the credit time that we want. For instance, if the user wants to create
a credit time to an employee because the employee need to have a credit
time each friday of the week. Then the user must select a calendar where
the working schedule is in 4/5 and there are no attendances for Friday.
With this way, the user has the possibility to change the credit time
by another type of credit time.

task-2314223