[FIX] project_timesheet_forecast: Fix the progress bar of timesheet
Currently, when role color is not defined, the progress bar is not
displayed becuase the default color is 0 and the class for that color
is not generated.

So in this commit, generate the class o_planning_progress_gantt_0 so
it should display the progressbar when no color is defined on role of
the planning.

closes odoo/enterprise#14828

Taskid: 2359038
X-original-commit: 5ddf5790fe33b0a70daca3048057793d588a3fc7
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>