[IMP] social_push_notifications: remove the need for 'firebase_admin' library
This commit removes the necessity of having the firebase_admin python library
installed to be able to send push notifications.

As the firebase_admin library is not officially supported in debian packages,
it was causing some deployment issues when trying to install the
social_push_notifications module.

We now allow having either the "firebase_admin" library installed OR simply the
"google_auth" library. As firebase_admin depends on google_auth, this should
work nicely for existing databases without needing to change anything.

The "google_auth" library is much lighter and most importantly officially
supported in debian packages. Which makes it a good replacement.

The implementation is also not so different from directly using the firebase
library, the important difference stands in the "return" type when sending
messages that is much more raw (POST request VS method call that returns
objects encapsulating the message result / batching / ...).

Task 2361171

closes odoo/enterprise#16691

X-original-commit: f71c2989b70983e3f573a652545ea23188776492
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>
Signed-off-by: awa-odoo <awa-odoo@users.noreply.github.com>