[IMP] l10n_mx_reports: use name_get on Mexico trial balance line names
On accounting reports, on the normal trial balance the name of the lines
are, gotten using account name_get method. The trial balance for Mexico defined
on l10n_mx_reports is using a direct string concatenation with code an
name of the account.

To be more consistent within the code, now trial balance for Mexico use
name_get too. This also allows to inherit name_get to change what the
report is showing on lines names if was necessary.

OPW#2375382

closes odoo/enterprise#14683

X-original-commit: 7eb62382b9906e74ac68045f533def6c7ba5d3fe
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>