[FIX] industry_fsm_report: allow fsm_user to edit existing Worksheet
- Go to Field Service > Configuration > Settings and enable "Worksheet Templates"
- Create 2 internal users with the following rights: (i.e. User X and User Y)
  * Sales: User: All Documents
  * Field Service: User
  * Project: User
- Go to Sales and create a Quotation with Product "Field Service"
- Save and Confirm (A task should be created)
- Connect with User X
- Go to Field Service and open the created task
- Click on Worksheet smart button and save a comment
- Connect with User Y
- Go to Field Service and open the created task
There is no indication on Worksheet smart button signaling that it is completed.
If User Y click on the smart button, he will not see previously entered comment
and will be able to enter his own comment.
But when trying to save it, the following error will be raised:
"duplicate key value violates unique constraint "x_project_worksheet_template_1_x_task_id_uniq"
DETAIL:  Key (x_task_id)=(42) already exists."

A Project user can edit a Worksheet, but he can only view the ones he has created.
So, when on a Task, he will not see if a Worksheet has already been done by someone
else, but will be able to create one.
As there can only be one Worksheet by Task, the unicity error is raised.

This commit will allow FSM users to view and edit all Worksheets.
This would prevent to force an user with the Project manager rights to allow him
to be able to view and edit an existing Worksheet.

opw-2368606

closes odoo/enterprise#14467

X-original-commit: 06587aa71d879f6186b05ff0675e00bdc45aeb3d
Signed-off-by: Anh Thao PHAM <kitan191@users.noreply.github.com>