[FIX] l10n_es_reports: remove taxes on all lines in report 349
The following commit https://github.com/odoo/enterprise/commit/852f6bcf9b0ad1e99013375a89eb63f0980fc67f was right about removing taxes, but it should be done
in all lines and not only for those having l10n_es_reports_mod349_invoice_type = 'A',
as we are already doing it for the summary section. We should also do it when computing
the total number of operations in the summary (for consistency only, as it is grouped
by partner anyway).

closes odoo/enterprise#14446

X-original-commit: 46bf0a4a22973c53b79273c46a19b9f11b7092f2
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Alex Tuyls <alt-odoo@users.noreply.github.com>