[FIX] web_grid: default week start/end for empty language
Problem
-------
user with no langage defined see only the anchor date
on grid view on week mode.

The method that get the start and the end of the day
depends on the user langage and does not fallback on a default
value

Solution
--------
If no langage is set on the user fallback on default begin and
end of the week as defined in the ISO 8601: monday as the begining
and sunday as the end.

closes odoo/enterprise#14457

X-original-commit: cc782c282898fbc2c9e13dba3ae1289808858f05
Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>