[FIX] stock_barcode: correct product for products with same lot name
- Install Sales, Inventory and Barcode
- Go to Inventory > Configuration > Settings and enable "Lots & Serial Numbers"
- Go to Inventory > Master Data > Products and create 2 Storable Products tracked by lots:
     Name     |   Barcode
  -----------------------
  Product A   |   AAAA
  Product B   |   BBBB

- Update Qty on hand to 100 for both Products and use the same lot name during the process (i.e. Lot 1)
- Go to Sales and create a Quotation with Product A and Product B
- Confirm
- A delivery is created
- Go to Barcode
- Scan "stock.picking" barcode
- Scan barcode of Product A => Product A will be highlighted
- Scan barcode of Lot 1 => Product A will be incremented
- Scan barcode of Product B => Product B will be highlighted
- Scan barcode of Lot 1 => Product A will be incremented instead of Product B

When scanning the barcode of a lot, a search is performed on the move lines to retrieve
the first line where the lot is used. The incremented Product is the one from this line.
However, in this case, several lines (Products) are associated to the scanned lot name.
Only retrieving the first found one is leading to this issue.

This commit will retrieve all move lines associated to the scanned lot.
If there are several lines, the chosen one will depend on the Product associated to the
last scanned line.

opw-2367966

closes odoo/enterprise#14496

X-original-commit: 217b3504146621657243078373ec688880e2dfae
Signed-off-by: William Henrotin <Whenrow@users.noreply.github.com>
Signed-off-by: Anh Thao PHAM <kitan191@users.noreply.github.com>