[FIX] mail_enterprise: Fix traceback for read property 'classList'
Currently, in expense module when we try to open line record it generates
traceback.
Step to reproduce:
 - Create an expense and write a description containing multiple line
(text fields).
 - Create a report
 - Try to open the line record

Issue is due to that when there is not chatter even it was going to
proces the attachment while rendering it.

So in this commit, We don't want to create a chatter when it is not
going to be displayed.

closes odoo/enterprise#14525

Taskid: 2365369
X-original-commit: e7db3c992045f96e46c0b90a4d7c7b7bea0b2f55
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>