[FIX] voip: fix traceback on missed call edge case
There's an edge case when the widget is open and someone calls then hangs up immediately
where the softphone of the receiver still thinks the call is ongoing.
When trying to reject or accept the call, the lib crashes.
This is a not so great fix that checks for that state of affairs and prevents the
acception/rejection from happening.

task-2326328

X-original-commit: d718ba35232063b06f3897ea4b0175063165b424