[FIX] web_enterprise: Home menu not usable anymore by screen reader users
The home menu was refactored in version 14.0 to handle all keyboard events
when searching, while keeping the focus on the search box. This allows to
input text natively on the text box, while being able to move over search
results at the same time, and the currently-selected result is highlighted
by toggling classes.

However, since the focus is always on the search box, screen reader users
are  not able anymore to know what result is being selected (in previous
versions, the focus actually moved to the selected element).

The above is fixed by labeling the currently-selected result using the
`aria-activedescendant` attribute, as well as adding required IDs and roles
for that attribute to work.

closes odoo/enterprise#14687

X-original-commit: b5fcb3a4a6c13200c25109a89e41936c683685bf
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>