[FIX] web_enterprise: toggle home menu when executing an action
Behavior prior to the commit:

When opening a dialog from the home menu (for example, the preference
dialog), the home menu stays displayed. Then, when opening a window
action from that dialog, the action can't display, since the home menu
shadows the web client.
For the actions that are triggered from the home menu itself, or the
dropdown menu, there is already an override (do_action) in the
Enterprise web client that turns off the home menu. But this override
does not get invoked for actions that are triggered from within a form.

Behavior after the commit:

The home menu is toggled off when invoking a window action, even when
this is not done directly under the web client.

Note:

Correction of commit cce561e568ef9a1ff947c3f43d3688aca10e59e3 which did
the override of current_action_updated incorrectly.

opw-2365539

closes odoo/enterprise#14639

X-original-commit: d005be6041fdb6d008472f2de9b5ff5c39487d39
Related: odoo/odoo#61491
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>