[IMP] hr_payroll: add sum_worked_days method
This method could be useful for a lot of given use cased.

Also add a method to compute the european leaves deduction. We dont rely
on the first implementation in hr_payroll as the use case is too complex.

In fact, when a employee takes a european leave, this leave is an advance on the
double holiday pay for the following year. That is, each european leave
taken reduce the double holiday pay in the following year.

task-2337215