[IMP] l10n_be_hr_payroll: add rule for european leaves deduction
This rule is added for the double holiday pay. It arrives when the
employee has taken some european leaves in previous working calendar.
In fact, this european leaves is an advance on the double holiday pay.
That is, each european leave taken reduces the double holiday pay for
the following year.

task-2337215