[FIX] sign: If value already exist then make field as readonly in sign
Purpose of the task is, when we have prefilled fields with form values
it shouldn't be editable once filled in.

It was working fine in 13.0 but when 'save' button is added by commit:
https://github.com/odoo/enterprise/commit/262ec8439e0afd2cb86b8b89c11cbcbddc4a0120#diff-6b29428415824ecd547de75f8a451b3b21b7dc74ac69a7364640034b90aaf840R344
it was removed because once someone save the value it should be editable
again when user open the same pdf second time.

So in this commit, If value is already prefilled or once saved it
will be in readonly mode it will not be changed again.

closes odoo/enterprise#14842

Taskid: 236554
X-original-commit: 55b775d0ce57403c9037db04884c06bae9a5a39c
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>