[FIX] web_studio: fix background image edition
In a multi-companies environment where company ids are not stored in cookies
and default company is selected upon connection, changing background image with
Studio will not display the new background immediately.
If the background had been changed previously, it will not be visible directly.
User will have to switch between companies to display the background image.

Also, resetting default background only works for user's default company.

The display issue comes from the fact that we are not checking the correct
company (the selected one) to know if there is a background image, but the
first one from "allowed_company_ids" set in context.

The solution to check background image of the correct company would be to use the
first company id retrieved from cookies, but only if user belongs to all retrieved
company ids.
Because when it is not the case, it means that connected user has changed and if
one of the companies is not available for him, the current company becomes user's
default one.
In all other cases, background image is checked from  user's default company.

opw-2371025

closes odoo/enterprise#14618

X-original-commit: 239c0fd9c0fbdb9dfaa08cdd58560f2d215b19d8
Signed-off-by: Anh Thao PHAM <kitan191@users.noreply.github.com>