[IMP] l10n_be_hr_payroll: update the calculation in credit time
When the employee asks a credit time or want to leave a credit time to
have a contract full-time, we need to recompute the amount of paid time
off.
For instance, if the employee has a contract full-time from 01/01/2019
and he asks a credit time to work at mid-time (5 days/week) from
02/01/2020, we need to recompute the amount of paid time off to
allocate him 20 days but 20 half days.
Because the employee works 5 morning/week, when he takes a leave, he
takes one entire day.

task-2314223