[FIX] account_reports: open journal items for non-range periods
When hitting the button "Journal Items" of an accounting report
with no "from date" boundary, such as the balance sheet,
a "from date" was nevertheless added to the domain,
therefore filtering out move lines older than the arbitrary date from set.

For instance, in the balance sheet,
when opening the journal items of an account,
all entries up to the end date must be displayed,
no date from should be set.

How to reproduce on runbot:
 1. Open the balance sheet,
 2. In the date filter, set "End of last month",
 3. Unfold "Bank And Cash accounts",
    then click on the first account and then click on "Journal Items"
 4. Observe there is no journal items shown,
    while the line in the report had data.

In `_query_get`, the method used to get the report amounts,
we can see account with `include_initial_balance` are taken into account
when `strict_range` is not enforced:
```
        if context.get('date_from'):
            if not context.get('strict_range'):
                domain += ['|', (date_field, '>=', context['date_from']), ('account_id.user_type_id.include_initial_balance', '=', True)]
            elif context.get('initial_bal'):
                domain += [(date_field, '<', context['date_from'])]
            else:
                domain += [(date_field, '>=', context['date_from'])]
```

Hence, it should be the case also when opening the journal items of an account
in such kind of reports.

The helper method `_get_options_date_domain`
from rev. 8cb9de9e6acb9bf17b29f813f277d987e0547f4b
seems actually the helper to call to handle this case.

opw-2337621

closes odoo/enterprise#14624

X-original-commit: c4997abef2fb7768edb04674810f7195dc879228
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Denis Ledoux (dle) <dle@odoo.com>