[FIX] l10n_be_hr*: replaces the use of work_time_rate in contract
The work time rate field in contract is not used to define a credit time.
Now, we need to use the work time rate in resource.calendar linked to
the contract.

That's why, we need to change the use of work_time_rate in contract by the
one in the calendar linked in the contract.

task-2314223