[REF] web_enterprise,web_mobile: simplify FieldMany2One on mobile
This commit simplifies the implementation of the FieldMany2One by avoiding
to override and reimplement the `_search()` function (which `search_val`
argument is only used when scanning a barcode).

Instead the search popup is directly opened when clicking the input and
the search is only performed when relevant (i.e. when scanning a
barcode).