[FIX] web_mobile: many2one_barcode on tablets
This commit is a cherry-pick of 32acb048f7149b529ee5fc00a626b927ea55e156
in which we added a test, as the change was already merge in this
branhc, only the test is forward-ported in this commit.

Before this commit, it was not possible to scan a barcode with the
camera on a tablet. It does work for mobile phones though.

This bug was present since we removed native dialogs:
See 6ca2d9e832a04df843b9aaeec54439f25f9cb006
_invokeMobileDialog was replaced by _search but the override of the latter
is only done for small screens, otherwise the file is escaped.
It's why it didn't worked for tablets (which have larger screens).

To fix this, we cherry-picked a commit from saas-14.1 where we use
_searchCreatePopup instead of using _search. It wasn't originally intended
to fix this issue but it does the job in an elegant way.
In addition, maintenance will also be easier by having common code between
version 14.0 and saas-14.1.

Steps to reproduce:
- Go to Field service
- Create a project.task
- Go on Studio and add a many2one_barcode product.product
- Open the camera by clicking on the widget
- After scanning, nothing is shown to the user.

opw-2466999

closes odoo/enterprise#17514

X-original-commit: bb3d4dd54b2404ba465053af6d682ccba64bf2c7
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>