[FIX] account_reports: include entries from archived journals in reports
Behavior prior to this commit:

When a journal is archived, its entries are still included in the
financial reports when "All" is selected under the journal filter.
However, the journal does not appear in the selection dropdown, and if
it was part of a journal group, selecting that journal group will not
include the related journal entries in the report.

Behavior after the commit:

The archived journal continues to be available in the financial report
(including when it is part of a journal group and the user selects that
group)

opw-2371035

closes odoo/enterprise#14645

X-original-commit: f033445fc99ec04a090e78b9ed90f1b5465a964c
Signed-off-by: Nicolas Galler <ngaller@users.noreply.github.com>