[IMP] account: more robustness in reconcilation query generation.
When getting the ir rules queries needed to be applied in
`_get_bank_statement_line_partners` a simple `replace` is used to
rename the table res_partner to p3. This can be an issue when
there are fields or other tables with the string 'res_partner'
within their names. The usage of a regex ensuring an exact match
it is a more robust way to do it.

closes odoo/enterprise#14652

X-original-commit: 4ec668c87b46571fc4e45c0c84af847ee7dcfe97
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>