[IMP] documents_spreasheet: pivot menu improvement
The current behaviour only allows a user to access the pivot properties
when selecting a pivot cell. This proves to be limitating as:
- there is no clear overview of every pivot included in the spreadsheet
- requesting a global data refresh needs to be done by selecting a cell
from each existing pivot in the spreadsheet which is painful

These can become confusing and tiresome as the number of pivots grows,
especially when working with several pivots on the same model.

This commit addresses the issue by
- adding an overview of all pivots accessible from both the side panel
and the top bar menu,
- adding the possibility to request a global measure refresh accessible
from a pivot property and the top bar menu.

task #2296980

closes odoo/enterprise#14697

Signed-off-by: Vincent Schippefilt (vsc) <vsc@odoo.com>