[FIX] l10n_mx_edi: return the cfdi file when an error is raise
Before this commit, just the message error was return, and not the
payment's cfdi document that generates the error. Not having the
document, complicates the understanding of the error and how to fix it.

Now, the generated payment's cfdi document as well as the error are
return.

Related to 06a82a94fc4310dcb9b0570717338891cfaaf33b

closes odoo/enterprise#14690

X-original-commit: f3571aed1ce12f05aa036ab0ab21d01dd5d123d5
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>