[REF] l10n_be_hr_payroll: compute complete months of work and presence prorata
To get the paid amount of the 13th and the double holiday pay, we need
to compute the number of complete months of work that the employee has
made and the presence prorata.

That's why, two new methods are created the one to compute the number
of complete months of work and the other one to compute the presence
prorata ratio of the employee.
And we use the both methods to get the paid amount for the 13th month
and the double holiday pay. Thus, we don't have duplicate code.

task-2337215