[IMP] social_{facebook,linkedin,twitter} : update module icons
This commit adds module icons for following modules:

- social_facebook
- social_linkedin
- social_twitter

Those are technical modules and we added icons which are customized from the
platforms' brand assets which follows the same design like other standard
modules / app icons for consistency.

However, few platforms are strict about using their brand assets as it is.
So, the icons we see while user tries to add a social stream are intact brand
assets because they are directly visible to the users (unlike module icons).

Task Id-2372492

closes odoo/enterprise#14685

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>