[IMP] sale_renting_crm: handle partner assignation on new rentals
Right now, from the opportunity form, user have an option create a new
rental quotation. But when the customer is not set on the opportunity,
it throws a UserError asking to set the customer first.

This commit improves the behavior by allowing user to create the customer
on the fly (the same way it is handles for the quotions with sale_crm
module).

TaskId - 2360863

closes odoo/enterprise#14689

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>