[FIX] l10n_mx_edi: payment with electronic transfer without partner bank
- Configure Mexican electronic invoicing;
- Create an invoice for Azure Interior;
- Make sure that Azure Interior don't have any bank assigned;
- Make sure the invoice is PPD by having payment terms with more than
  one line assigned on it;
- Validate and 'Send Now' to get a Folio Fiscal / CFDI number;
- Register a payment, with 'Transferencia electrónica de fondos' as
  Payment Way;
- Go to the payment and 'Send Now' to get the CFDI number;

Before this commit, an error was raised: 'El campo NomBancoOrdExt se
debe registrar'.

This issue occurs because the field RfcEmisorCtaOrd (emitter_vat_ord)
was filled with the vat : "XEXX010101000". The vat "XEXX010101000" is a
generic value to be use as vat when and only when the user is not
Mexican, and in that case the NomBancoOrdExt must be filled.

opw-2363916
opw-2378274

closes odoo/enterprise#14703

X-original-commit: 02f5221181d25917596ce1d2655d781a822abc8d
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>