[ADD] l10n_pe_edi: Peruvian invoice EDI
From the manifest:
Allow the user to generate the EDI document for Peruvian invoicing.

By default, the system uses the IAP proxy.  This has the advantage that you
can use the system immediately the moment you choose Digiflow as your OSE
in the SUNAT portal.

You can also directly send it to Digiflow if you bought an account from them
and even to SUNAT in case of contingency.

We support sending and cancelling of customer invoices.

closes odoo/enterprise#16058

X-original-commit: 3dc0c43f5f5590b335db7477bc3aab8f6483e389
Related: odoo/odoo#65341
Signed-off-by: Josse Colpaert <jco@openerp.com>